filter selector problem solved
when the list changes in the combobox, the first item changes too, which is the default. It means that the current text is changing too. The solution: I use the generated handler function for connecting the combobox to the main handler function only. With this way, the first change won't trigger the main handler